[add backend float support] Test SysY compiler 2023 functional: Total: 98, Passed: 67, Failed: 31 PASSED: 67 RUN_COMPILER_FAILED: 2 LINK_EXECUTABLE_FAILED: 8 RUN_EXECUTABLE_FAILED: 0 OUTPUT_MISMATCH: 21

+++ src/mir/mir.cpp
@@ -31,7 +31,7 @@ void MIRZeroStorage::print(std::ostream& os, CodeGenContext& ctx) {}
 void MIRDataStorage::print(std::ostream& os, CodeGenContext& ctx) {
     for (auto& val : _data) {
         os << "\t.4byte\t";
-        if (is_float()) os << (float)val << std::endl;
+        if (is_float()) os << val << std::endl;
         else os << val << std::endl;
     }
 }
